test(#361): move read simple-script test from file to dedicated module
